refactor(lsp): modularize server.rs into server/ submodules
Pre-commit hook ran. Total eslint: 0, total circular: 0
